Fix query stats: wire SamplesFetched/BytesFetched to storage fetch
Cubic P2 was valid: addSamplesFetched/addBytesFetched existed but
had no callers, so maybeLogQueryStats always logged samples_fetched=0
bytes=0 contradicting docs/docs/query-stats.md and
dashboards/query-stats.json which expect real values since v1.116.0.

Root cause: netstorage.ProcessSearchQuery computed samples
(sum br.RowsCount) and bytes (tbf.Len) internally for
maxSamplesPerQuery and qt.Printf but discarded them on return via
Results{sr,tbf,pts}. evalRollupFuncNoCache only called
addSeriesFetched(rss.Len()).

Fix with minimal propagation reusing existing structures:
- netstorage.Results: add samples int / bytes uint64 fields and
  getters SamplesFetched()/BytesFetched(); store them in
  ProcessSearchQuery (rss.samples=samples; rss.bytes=tbf.Len()).
- promql.QueryStats: change addBytesFetched to uint64 (match tbf.Len)
  and wire in eval.go after storage fetch:
  qs.addSamplesFetched(rss.SamplesFetched())
  qs.addBytesFetched(rss.BytesFetched())
  Handles nil qs and concurrent accumulates (atomic).
- Docs/dashboard already expect samples_fetched (raw samples) and
  bytes (storage wire, tbf.Len) – now populated correctly; no
  misleading zeros or incorrect semantics.

Tests: extend TestQueryStats_NilSafety for new setters, add
TestQueryStats_SamplesAndBytesFetched covering initial 0,
accumulation, concurrent adds and independence.

+++ app/vmselect/promql/query_stats.go
@@ -64,7 +64,7 @@ func (qs *QueryStats) addSamplesFetched(n int) {
 	qs.SamplesFetched.Add(int64(n))
 }
 
-func (qs *QueryStats) addBytesFetched(n int) {
+func (qs *QueryStats) addBytesFetched(n uint64) {
 	if qs == nil {
 		return
 	}
